requirements-ci.txt
# Requisitos otimizados para CI (GitHub Actions)
watchdog
pyyaml
pypdf
pytesseract
pillow
joblib
scikit-learn
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
psutil>=5.9.0
typer
rich
pytest
pytest-mock
pytest-cov
pytest-timeout
flake8==7.0.0
black==24.1.1
isort==5.13.2
mypy
bandit
safety
types-PyYAML
types-psutil
pre-commit

# Inteligência de Vídeo e Áudio para CI
openai-whisper>=20230314
librosa>=0.10.0
soundfile>=0.12.0
